+++ src/main/java/org/cptgum/philosophersstone/PhilosophersStone.java
@@ -1,14 +1,17 @@
 package org.cptgum.philosophersstone;
 
+import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
+import org.cptgum.philosophersstone.listener.PlayerInteractListener;
 import org.cptgum.philosophersstone.utils.Translator;
 import org.cptgum.philosophersstone.utils.ConfigLoader;
 import org.cptgum.philosophersstone.commands.Commands;
 import org.cptgum.philosophersstone.utils.LoggerUtils;
 
-public class PhilosophersStone extends JavaPlugin {
-        private final ConfigLoader configLoader;
+import java.util.HashMap;
 
+public class PhilosophersStone extends JavaPlugin {
+    private final ConfigLoader configLoader;
     public PhilosophersStone() {
         configLoader = new ConfigLoader(this);
     }
@@ -18,16 +21,18 @@ public class PhilosophersStone extends JavaPlugin {
         // Plugin startup logic
         // Load config
         configLoader.loadConfig();
+
         // Load languages
         Translator translator = Translator.getInstance();
         translator.loadLanguages(this);
+
+        //Register listeners
+        Bukkit.getServer().getPluginManager().registerEvents(new PlayerInteractListener(), this);
 
         //If possible, initialize last:
         // Register commands
         getCommand("philosophersstone").setExecutor(new Commands());
         LoggerUtils.logInfo("Plugin started!");
-        //Example Usage of the Translator
-        LoggerUtils.logInfo(translator.translate("examplemain.examplesub"));
     }